requirements.txt
Django==3.2
djangorestframework==3.13.1
drf-spectacular==0.22.1
black==22.6.0
django-debug-toolbar==3.5.0
psycopg2==2.9.3
django-phonenumber-field
django-phonenumbers
Pillow==9.2.0
djoser==2.1.0
django-ckeditor==6.4.2